refactor: split service_test.go into service_list_test.go
